requirements.txt
aws-cdk-lib>=2.76.0
constructs>=10.0.0
boto3>=1.26.0
requests>=2.28.0
pandas>=1.5.0
matplotlib>=3.6.0
itables>=1.5.0
jupyter>=1.0.0
tabulate>=0.9.0